Extract Networks to constants
