+++ packages/extension/src/chromeServices/background.ts
@@ -52,7 +52,7 @@ chrome.runtime.onMessage.addListener((message: BackgroundMessage, sender, sendRe
   // We make sure that the message comes from gem-wallet
   if (app === GEM_WALLET && sender.id === chrome.runtime.id) {
     if (type === Message.RequestNetwork) {
-      sendResponse(Network.Test);
+      sendResponse(Network.TESTNET);
     } else if (type === Message.RequestAddress) {
       chrome.windows.getAll().then((openedWindows) => {
         // We check if the popup is currently open
