+++ packages/extension/src/constants/index.ts
@@ -2,7 +2,6 @@ export * from './colors';
 export * from './links';
 export * from './localStorage';
 export * from './navigation';
-export * from './network';
 export * from './parameters';
 export * from './routes';
 export * from './sizes';
