+++ packages/extension/src/contexts/NetworkContext/NetworkContext.tsx
@@ -1,9 +1,8 @@
 import { useContext, useState, useEffect, createContext, FC, useCallback } from 'react';
 import * as Sentry from '@sentry/react';
+import { NETWORK, Network } from '@gemwallet/constants';
 import { Client } from 'xrpl';
-import { Network } from '../../types';
 import { loadNetwork, removeNetwork, saveNetwork } from '../../utils';
-import { NETWORK } from '../../constants';
 
 interface ContextType {
   switchNetwork: (network: Network) => void;
