+++ packages/extension/src/types/index.ts
@@ -1,4 +1,3 @@
-export * from './network.types';
 export * from './transaction.types';
 export * from './utils.types';
 export * from './wallet.types';
