--- packages/extension/src/types/network.types.ts
@@ -1,13 +1,0 @@
-export enum Network {
-  MAINNET = 'Mainnet',
-  TESTNET = 'Testnet',
-  DEVNET = 'Devnet',
-  NFT_DEVNET = 'NFTDevnet'
-}
-
-export enum NetworkServer {
-  MAINNET = 'wss://xrplcluster.com',
-  TESTNET = 'wss://s.altnet.rippletest.net:51233',
-  DEVNET = 'wss://s.devnet.rippletest.net:51233',
-  NFT_DEVNET = 'wss://xls20-sandbox.rippletest.net:51233'
-}
